requirements.txt
streamlit==1.31.1

numpy==1.26.4
pandas==2.1.4

torch==2.2.2
torchvision==0.17.2

pillow==10.2.0
scikit-learn==1.4.2
seaborn==0.13.2
matplotlib==3.8.4

scipy
openpyxl
